extract getTavernDescription
